:white_check_mark: test: add initial pytest suite
