+++ terraform/modules/content_delivery_network/lambda_functions/redirect_s3_404/src/lambda_function.py
@@ -64,40 +64,3 @@ def replace_implementation(path_parts):
     implementation = path_parts[impl_path_location]
     path_parts[impl_path_location] = "dynamic"
     return implementation
-
-
-def test_handler():
-    # Mock event and context
-    event = {
-        "Records": [
-            {
-                "cf": {
-                    "response": {
-                        "status": "404",
-                        "headers": {
-                            "content-type": [
-                                {"key": "Content-Type", "value": "text/plain"}
-                            ]
-                        },
-                    },
-                    "request": {
-                        "uri": "/sbtn_natural_forests_map/v202310/natural_forest/10/20/30.png",
-                        "querystring": "some_param=30",
-                    },
-                }
-            }
-        ]
-    }
-    context = {}
-
-    # Call the handler function
-    response = handler(event, context)
-
-    # Assertions
-    assert response["status"] == "307"
-    assert response["headers"]["location"][0]["key"] == "Location"
-    assert (
-        response["headers"]["location"][0]["value"]
-        == "/sbtn_natural_forests_map/v202310/dynamic/10/20/30.png?some_param=30&implementation=natural_forest"
-    )
-    assert response["headers"]["content-type"][0]["value"] == "application/json"
